pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pilot_academy"
version = "0.1.0"
description = "Pilot Academy - TensorFlow GPU project"
requires-python = ">=3.8"
dependencies = [
    "tensorflow[and-cuda]",
    "numpy",
    "pandas",
    "matplotlib",
    "tqdm",
    "opencv-python",
    "wandb",
    "python-dotenv",
    "hydra-core"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]